int -> ulong

+++ language-testing/fibonacci/fibonacci.d
@@ -1,9 +1,9 @@
 import std.stdio;
 import std.conv;
 
-int fib(int n) {
-  int old_value = 1;
-  int new_value = 0;
+ulong fib(ulong n) {
+  ulong old_value = 1;
+  ulong new_value = 0;
 
   foreach (_; 0..n) {
     immutable auto tmp = new_value;
@@ -17,14 +17,14 @@ int fib(int n) {
 void main(string[] args) {
   if (args.length > 1) {
     try {
-      writeln(fib(to!int(args[1])));
+      writeln(fib(to!ulong(args[1])));
     } catch (ConvException e) {
       writefln("%s: Failed to convert %s to digit", args[0], args[1]);
     }
 
   } else {
     foreach (i; 0..10) {
-      writefln("%d\t%d", fib(i), fib(i+10));
+      writefln("%lu\t%lu", fib(i), fib(i+10));
     }
   }
 }
